Puzzle11, Part2

+++ src/main/kotlin/puzzle11/Monkey.kt
@@ -2,7 +2,7 @@ package puzzle11
 
 data class Monkey(
     val monkeyAsString: String,
-    val items: MutableList<Int>,
+    val items: MutableList<ULong>,
     val operationOnItem: OperationOnItem,
     val testOnItem: TestOnItem,
 )
