+++ src/main/kotlin/puzzle11/OperationOnItem.kt
@@ -2,5 +2,5 @@ package puzzle11
 
 data class OperationOnItem(
     val operationString: String,
-    val operation: (Int) -> Int,
+    val operation: (ULong) -> ULong,
 )
